autoscaler/requirements.txt
kubernetes
pyyaml
prometheus-api-client
prometheus_client
requests
